+++ nft-ipfs/scripts/connect_to_pinata.py
@@ -1,13 +1,14 @@
 """
 Code modified from: https://github.com/Vourhey/pinatapy
 """
-from unicodedata import name
-import requests
-import os
 import json
-import typing as tp
+import os
 import time
+import typing as tp
 from pathlib import Path
+from unicodedata import name
+
+import requests
 from dotenv import load_dotenv
 
 load_dotenv()
